Add flush after header sent
This makes it possible to send the headers before the content is sent to the
client, so they can already be processed and evaluated before the first results
show up.

This is especially interesting when having big streamed responses, like a csv
report. It also makes it possible to implement one shot server sent events.

Signed-off-by: Marius Büscher <marius@gridscale.io>